Restraunt
A restaurant is taking online orders for lunch. Assume lunch orders are to be filled in the arrangement in which they are received, unless they have a priority code. Write a program that receives the order, four-digit personal code and the number of lunch orders requested, and stores this information in a transaction list. If the code exists in the priority list, use an algorithm of your choice to determine the position of this order on the transaction list. The program should end with a display of the transaction list
(Menu Item, price, id and number of lunches). Assume priority codes are numbers between 1 and 500.

priority queue